+++ R/account_account.R
@@ -58,6 +58,7 @@ pkg.accounts <- new.env()   # account information that we've fetched.
 #' @seealso [topic_trees()] to find the topic trees in an account.
 #' @seealso [logs()] to read the account's logs.
 #' @seealso [users()] to see what users have access to the account.
+#' @seealso [rules()] for the list of rules applied to mentions in the account.
 #'
 #' @export
 account <- function(codes, ...) {